refactor: moved load-script and load-stylesheet to lib directory
